refactor: refactor to Clear architecture
