--- app/src/main/java/com/example/multimathsolver/data/booleanalgebra/BooleanFunction.java
@@ -1,400 +1,0 @@
-package com.example.multimathsolver.data.booleanalgebra;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-
-public class BooleanFunction {
-    private final List<String> parameters;
-    private final int[][] table;
-    public BooleanFunction(ExpressionHandlerInterface expressionHandler) throws IncorrectFunctionInput {
-        this.parameters = expressionHandler.getExpressionParameters();
-        this.table = fillTheTable();
-        boolean isException = expressionHandler.fillTableWithValueOfFunction(this.table);
-        if (isException) {
-            throw new IncorrectFunctionInput("Неверно введена функция");
-        }
-    }
-
-    public List<String> getListOfParameters() {
-        return new ArrayList<>(parameters);
-    }
-
-    private int[][] fillTheTable() {
-        int[][] table = new int[(int) Math.pow(2, parameters.size())][parameters.size() + 1];
-        int x = (int) Math.pow(2, parameters.size());
-        for(int col = 0; x != 1; ++col) {
-            x /= 2;
-            boolean res = true;
-            for(int i = 0; i < table.length; ++i) {
-                if (i % x == 0) {
-                    res = !res;
-                }
-                table[i][col] = res ? 1 : 0;
-            }
-        }
-        return table;
-    }
-
-    public int[][] getTable() {
-        int[][] output = new int[table.length][table[0].length];
-        for (int i = 0; i < table.length; ++i) {
-            System.arraycopy(table[i], 0, output[i], 0, table[0].length);
-        }
-        return output;
-    }
-
-
-    public String getPerfectDNF() {
-        StringBuilder output = new StringBuilder();
-        for (int[] ints : table) {
-            if (ints[table[0].length - 1] == 1) {
-                output.append("(");
-                for (int j = 0; j < table[0].length - 1; ++j) {
-                    if (ints[j] == 0) {
-                        output.append(UnaryOperation.NEGATION.charOfOperation);
-                    }
-                    output.append(parameters.get(j))
-                            .append(" ")
-                            .append(BinaryOperation.CONJUNCTION.charOfOperation)
-                            .append(" ");
-                }
-                if (output.length() >= 3) {
-                    output = new StringBuilder(output.substring(0, output.length() - 3));
-                }
-                output.append(") ").append(BinaryOperation.DISJUNCTION.charOfOperation).append(" ");
-            }
-        }
-        if (output.length() >= 3) {
-            output = new StringBuilder(output.substring(0, output.length() - 3));
-        }
-        return output.toString();
-    }
-
-    public String getPerfectCNF() {
-        StringBuilder output = new StringBuilder();
-        for (int[] ints : table) {
-            if (ints[table[0].length - 1] == 0) {
-                output.append("(");
-                for (int j = 0; j < table[0].length - 1; ++j) {
-                    if (ints[j] == 1) {
-                        output.append(UnaryOperation.NEGATION.charOfOperation);
-                    }
-                    output.append(parameters.get(j))
-                            .append(" ")
-                            .append(BinaryOperation.DISJUNCTION.charOfOperation)
-                            .append(" ");
-                }
-                if (output.length() >= 3) {
-                    output = new StringBuilder(output.substring(0, output.length() - 3));
-                }
-                output.append(") ").append(BinaryOperation.CONJUNCTION.charOfOperation).append(" ");
-            }
-        }
-        if (output.length() >= 3) {
-            output = new StringBuilder(output.substring(0, output.length() - 3));
-        }
-        return output.toString();
-    }
-
-    private List<List<Integer>> getListOfSDNF() {
-        List<List<Integer>> listOfAbbreviatedDNF = new ArrayList<>();
-        for (int[] row : table) {
-            if (row[row.length - 1] == 1) {
-                List<Integer> listOfValuesOfParameters = new ArrayList<>();
-                for (int j = 0; j < row.length - 1; ++j) {
-                    listOfValuesOfParameters.add(row[j]);
-                }
-                listOfAbbreviatedDNF.add(listOfValuesOfParameters);
-            }
-        }
-        return listOfAbbreviatedDNF;
-    }
-
-    private List<List<Integer>> getListOfAbbreviatedDNF() {
-        List<List<Integer>> abbreviatedDNF = getListOfSDNF();
-        do {
-            List<List<Integer>> clone = new ArrayList<>(abbreviatedDNF);
-            int counter = 0;
-            for (int k = 0; k < parameters.size(); ++k) {
-                for (int j = 0; j < abbreviatedDNF.size(); ++j) {
-                    for (int i = j + 1; i < abbreviatedDNF.size(); ++i) {
-                        if (!abbreviatedDNF.get(j).get(k).equals(abbreviatedDNF.get(i).get(k))) {
-                            List<Integer> lst1 = new ArrayList<>(abbreviatedDNF.get(j));
-                            List<Integer> lst2 = new ArrayList<>(abbreviatedDNF.get(i));
-                            lst1.set(k, -1);
-                            lst2.set(k, -1);
-                            if (lst1.equals(lst2)) {
-                                clone.add(lst1);
-                                counter++;
-                                clone.remove(abbreviatedDNF.get(i));
-                                clone.remove(abbreviatedDNF.get(j));
-                            }
-                        }
-                    }
-                }
-            }
-            Set<List<Integer>> setClone = new HashSet<>(clone);
-            abbreviatedDNF = new ArrayList<>(setClone);
-            if (counter == 0) break;
-        }
-        while (true);
-        return abbreviatedDNF;
-    }
-
-    public String getAbbreviatedDNF() {
-        List<List<Integer>> listOfAbbreviatedDNF = getListOfAbbreviatedDNF();
-        StringBuilder output = new StringBuilder();
-        for (List<Integer> conjunction : listOfAbbreviatedDNF) {
-            output.append("(");
-            for (int j = 0; j < conjunction.size(); ++j) {
-                Integer el = conjunction.get(j);
-                if (el != -1) output
-                        .append(el == 1 ? parameters.get(j) : UnaryOperation.NEGATION.charOfOperation + parameters.get(j))
-                        .append(" ").append(BinaryOperation.CONJUNCTION.charOfOperation)
-                        .append(" ");
-            }
-            if (output.length() >= 3) {
-                output = new StringBuilder(output.substring(0, output.length() - 3));
-            }
-            output.append(") ")
-                    .append(BinaryOperation.DISJUNCTION.charOfOperation)
-                    .append(" ");
-
-        }
-        if (output.length() >= 3) {
-            output = new StringBuilder(output.substring(0, output.length() - 3));
-        }
-        return output.toString();
-    }
-
-    private List<List<List<Integer>>> getListOfDeadLockedDNF() {
-        List<List<List<Integer>>> tupic = new ArrayList<>();
-        tupic.add(getListOfAbbreviatedDNF());
-        int j = 0;
-        while (j != tupic.size()) {
-            List<List<Integer>> list = tupic.get(j);
-            for (int i = 0; i < list.size(); ++i) {
-                List<List<Integer>> copy = new ArrayList<>(list);
-                //noinspection SuspiciousListRemoveInLoop
-                copy.remove(i);
-                if (isDNF(copy)) tupic.add(copy);
-            }
-            ++j;
-        }
-        tupic = cleanDeadLockedDNF(tupic);
-        return tupic;
-    }
-
-    private boolean isDNF(List<List<Integer>> list) {
-        for (int[] row : table) {
-            if (row[row.length - 1] == 1) {
-                int result = 0;
-                for (List<Integer> integers : list) {
-                    boolean flag = true;
-                    for (int j = 0; j < integers.size(); ++j) {
-                        if (integers.get(j) != -1) {
-                            if (integers.get(j) != row[j]) {
-                                flag = false;
-                                break;
-                            }
-                        }
-                    }
-                    if (flag) ++result;
-                }
-                if (result == 0) return false;
-            }
-        }
-        return true;
-    }
-
-    private List<List<List<Integer>>> cleanDeadLockedDNF(List<List<List<Integer>>> list) {
-        List<List<List<Integer>>> copy = new ArrayList<>(list);
-        List<Integer> removable = new ArrayList<>();
-        for (int i = list.size() - 1; i > 0; --i) {
-            for (int j = 0; j < i; j++) {
-                List<List<Integer>> lst1 = list.get(i);
-                List<List<Integer>> lst2 = list.get(j);
-                boolean res = true;
-                for (List<Integer> vars : lst1) {
-                    if (!lst2.contains(vars)) {
-                        res = false;
-                        break;
-                    }
-                }
-                if (res) {
-                    removable.add(j);
-                }
-            }
-        }
-        Set<Integer> set = new HashSet<>(removable);
-        removable.clear();
-        removable.addAll(set);
-        Collections.sort(removable);
-        Collections.reverse(removable);
-        for (Integer index : removable) copy.remove((int) index);
-        return copy;
-    }
-
-    public String getDeadLockedDNF() {
-        StringBuilder output = new StringBuilder();
-        for (List<List<Integer>> dnf : getListOfDeadLockedDNF()) {
-            for (List<Integer> conjunction : dnf) {
-                output.append("(");
-                for (int i = 0; i < conjunction.size(); ++i) {
-                    if (conjunction.get(i) != -1)
-                        output.append(conjunction.get(i) == 1
-                                ? parameters.get(i)
-                                : UnaryOperation.NEGATION.charOfOperation + parameters.get(i))
-                                .append(" ").append(BinaryOperation.CONJUNCTION.charOfOperation)
-                                .append(" ");
-                }
-                if (output.length() >= 3) {
-                    output = new StringBuilder(output.substring(0, output.length() - 3));
-                }
-                output.append(") ").append(BinaryOperation.DISJUNCTION.charOfOperation).append(" ");
-            }
-            if (output.length() >= 3) {
-                output = new StringBuilder(output.substring(0, output.length() - 3));
-            }
-            output.append("\n");
-        }
-        if (output.length() >= 1) output = new StringBuilder(output.substring(0, output.length()-1));
-        return output.toString();
-    }
-
-    private List<List<List<Integer>>> getListOfMinimalDNF() {
-        int minimum = Integer.MAX_VALUE;
-        List<List<List<Integer>>> listOfMinimalDNF = new ArrayList<>();
-        List<List<List<Integer>>> deadLockedDNF = getListOfDeadLockedDNF();
-        for (List<List<Integer>> dnf : deadLockedDNF) {
-            int counter = 0;
-            for (List<Integer> conjunction : dnf) {
-                for (Integer literal : conjunction) {
-                    if (literal != -1) ++counter;
-                }
-            }
-            minimum = Math.min(counter, minimum);
-        }
-        for (List<List<Integer>> dnf : deadLockedDNF) {
-            int counter = 0;
-            for (List<Integer> conjunction : dnf) {
-                for (Integer literal : conjunction) {
-                    if (literal != -1) ++counter;
-                }
-            }
-            if (counter == minimum) listOfMinimalDNF.add(dnf);
-        }
-        return listOfMinimalDNF;
-    }
-
-    public String getMinimalDNF() {
-        StringBuilder output = new StringBuilder();
-        for (List<List<Integer>> dnf : getListOfMinimalDNF()) {
-            for (List<Integer> conjunction : dnf) {
-                output.append("(");
-                for (int i = 0; i < conjunction.size(); ++i) {
-                    if (conjunction.get(i) != -1)
-                        output.append((conjunction.get(i) == 1 ? parameters.get(i) :
-                                UnaryOperation.NEGATION.charOfOperation + parameters.get(i)))
-                                .append(" ").append(BinaryOperation.CONJUNCTION.charOfOperation)
-                                .append(" ");
-                }
-                if (output.length() >= 3) output = new StringBuilder(output.substring(0, output.length() - 3));
-                output.append(") ").append(BinaryOperation.DISJUNCTION.charOfOperation).append(" ");
-            }
-            if (output.length() >= 3) output = new StringBuilder(output.substring(0, output.length() - 3));
-            output.append("\n");
-        }
-        if (output.length() >= 1) output = new StringBuilder(output.substring(0, output.length() - 1));
-        return output.toString();
-    }
-
-    private List<Integer> mergeLists(List<Integer> arr1, List<Integer> arr2) {
-        List<Integer> res = new ArrayList<>(arr1);
-        for (int i = 0; i < arr2.size(); ++i) {
-            res.add((arr1.get(i) + arr2.get(i)) % 2);
-        }
-        return res;
-    }
-
-    public String getPolynomial() {
-        StringBuilder output = new StringBuilder();
-        List<List<Integer>> entry = new ArrayList<>();
-        for (int[] ints : table) {
-            List<Integer> valueOfFunction = new ArrayList<>();
-            valueOfFunction.add(ints[parameters.size()]);
-            entry.add(valueOfFunction);
-        }
-        while (entry.size() != 1) {
-            List<List<Integer>> clone = new ArrayList<>();
-            for (int i = 0; i < entry.size() - 1; i += 2) {
-                clone.add(mergeLists(entry.get(i), entry.get(i + 1)));
-            }
-            entry = clone;
-        }
-        List<Integer> result = entry.get(0);
-        for (int i = table.length - 1; i > 0; --i) {
-            if (result.get(i) == 1) {
-                for (int j = 0; j < table[i].length - 1; ++j) {
-                    if (table[i][j] == 1) output.append(parameters.get(j));
-                }
-                output.append(" ").append(BinaryOperation.XOR.charOfOperation).append(" ");
-            }
-        }
-        if (result.get(0) == 1) output.append(1);
-        else {
-            if (output.length() >= 3) output = new StringBuilder(output.substring(0, output.length() - 3));
-        }
-        return output.toString();
-    }
-
-    private boolean belongsToT0() {
-        return table[0][parameters.size()] == 0;
-    }
-
-    private boolean belongsToT1() {
-        return table[table.length-1][parameters.size()] == 1;
-    }
-
-    private boolean belongsToL() {
-        for (String string : getPolynomial().split(" ")) {
-            if (!string.equals(BinaryOperation.XOR.charOfOperation)) {
-                if (!parameters.contains(string)) return false;
-            }
-        }
-        return true;
-    }
-
-    private boolean belongsToS() {
-        for (int i = 0; i < table.length / 2; ++i) {
-            if (table[i][parameters.size()] == table[table.length - 1 - i][parameters.size()]) return false;
-        }
-        return true;
-    }
-
-    private boolean belongsToM(int[] arr) {
-        if (arr.length <= 1) return true;
-        for (int i = 0; i < arr.length / 2; ++i) {
-            if (arr[i] > arr[i + (arr.length / 2)]) return false;
-        }
-        return belongsToM(Arrays.copyOfRange(arr, 0, arr.length / 2)) && belongsToM(Arrays.copyOfRange(arr, arr.length / 2, arr.length));
-    }
-
-    public Map<PostClass, Boolean> getBelongingToPostClasses() {
-        int[] functionValue = new int[table.length];
-        for (int i = 0 ; i < table.length; ++i) functionValue[i] = table[i][parameters.size()];
-        Map<PostClass, Boolean> map = new HashMap<>();
-        map.put(PostClass.T0, belongsToT0());
-        map.put(PostClass.T1, belongsToT1());
-        map.put(PostClass.L, belongsToL());
-        map.put(PostClass.M, belongsToM(functionValue));
-        map.put(PostClass.S, belongsToS());
-        return map;
-    }
-}
