--- app/src/main/java/com/example/multimathsolver/data/booleanalgebra/ExpressionHandlerInterface.java
@@ -1,8 +1,0 @@
-package com.example.multimathsolver.data.booleanalgebra;
-
-import java.util.List;
-
-public interface ExpressionHandlerInterface {
-    List<String> getExpressionParameters();
-    boolean fillTableWithValueOfFunction(int[][] table);
-}
