+++ app/src/main/java/com/example/multimathsolver/domain/GetAbbreviatedDNFUseCase.java
@@ -1,7 +1,5 @@
 package com.example.multimathsolver.domain;
 
-import com.example.multimathsolver.data.booleanalgebra.BooleanFunction;
-
 public class GetAbbreviatedDNFUseCase {
     private final Repository repository;
 
@@ -9,7 +7,7 @@ public class GetAbbreviatedDNFUseCase {
         this.repository = repository;
     }
 
-    public String getAbbreviatedDNFUseCase(BooleanFunction b) {
-        return repository.getAbbreviatedDNF(b);
+    public String getAbbreviatedDNFUseCase(BooleanFunction function) {
+        return repository.getAbbreviatedDNF(function);
     }
 }
