+++ app/src/main/java/com/example/multimathsolver/domain/GetBooleanFunctionUseCase.java
@@ -1,8 +1,5 @@
 package com.example.multimathsolver.domain;
 
-import com.example.multimathsolver.data.booleanalgebra.BooleanFunction;
-import com.example.multimathsolver.data.booleanalgebra.ExpressionHandlerInterface;
-import com.example.multimathsolver.data.booleanalgebra.IncorrectFunctionInput;
 
 public class GetBooleanFunctionUseCase {
     private final Repository repository;
@@ -11,7 +8,7 @@ public class GetBooleanFunctionUseCase {
         this.repository = repository;
     }
 
-    public BooleanFunction getBooleanFunctionUseCase(ExpressionHandlerInterface expressionHandler) throws IncorrectFunctionInput {
-        return repository.getBooleanFunction(expressionHandler);
+    public BooleanFunction getBooleanFunctionUseCase(String expression) throws IncorrectFunctionInput {
+        return repository.getBooleanFunction(expression);
     }
 }
