+++ app/src/main/java/com/example/multimathsolver/domain/GetPerfectCNFUseCase.java
@@ -1,7 +1,5 @@
 package com.example.multimathsolver.domain;
 
-import com.example.multimathsolver.data.booleanalgebra.BooleanFunction;
-
 public class GetPerfectCNFUseCase {
     private final Repository repository;
 
@@ -9,7 +7,7 @@ public class GetPerfectCNFUseCase {
         this.repository = repository;
     }
 
-    public String getPerfectCNFUseCase(BooleanFunction b) {
-        return repository.getPerfectCNF(b);
+    public String getPerfectCNFUseCase(BooleanFunction function) {
+        return repository.getPerfectCNF(function);
     }
 }
