+++ app/src/main/java/com/example/multimathsolver/domain/GetPolynomialUseCase.java
@@ -1,7 +1,5 @@
 package com.example.multimathsolver.domain;
 
-import com.example.multimathsolver.data.booleanalgebra.BooleanFunction;
-
 public class GetPolynomialUseCase {
     private final Repository repository;
 
@@ -9,7 +7,7 @@ public class GetPolynomialUseCase {
         this.repository = repository;
     }
 
-    public String getPolynomialUseCase(BooleanFunction b) {
-        return repository.getPolynomial(b);
+    public String getPolynomialUseCase(BooleanFunction function) {
+        return repository.getPolynomial(function);
     }
 }
