+++ app/src/main/java/com/example/multimathsolver/domain/IncorrectFunctionInput.java
@@ -1,4 +1,4 @@
-package com.example.multimathsolver.data.booleanalgebra;
+package com.example.multimathsolver.domain;
 
 import java.io.IOException;
