+++ app/src/main/java/com/example/multimathsolver/domain/Repository.java
@@ -1,14 +1,9 @@
 package com.example.multimathsolver.domain;
 
-import com.example.multimathsolver.data.booleanalgebra.BooleanFunction;
-import com.example.multimathsolver.data.booleanalgebra.PostClass;
-import com.example.multimathsolver.data.booleanalgebra.ExpressionHandlerInterface;
-import com.example.multimathsolver.data.booleanalgebra.IncorrectFunctionInput;
-
 import java.util.Map;
 
 public interface Repository {
-    BooleanFunction getBooleanFunction(ExpressionHandlerInterface expressionHandler) throws IncorrectFunctionInput;
+    BooleanFunction getBooleanFunction(String expression) throws IncorrectFunctionInput;
 
     String getPerfectDNF(BooleanFunction function);
