+++ app/src/test/java/com/example/multimathsolver/booleanalgebra/DeadLockedDNFTest.java
@@ -1,8 +1,9 @@
 package com.example.multimathsolver.booleanalgebra;
 
-import com.example.multimathsolver.data.booleanalgebra.BooleanFunction;
-import com.example.multimathsolver.data.booleanalgebra.IncorrectFunctionInput;
-import com.example.multimathsolver.data.booleanalgebra.StringExpressionHandler;
+import com.example.multimathsolver.data.RepositoryImpl;
+import com.example.multimathsolver.domain.BooleanFunction;
+import com.example.multimathsolver.domain.IncorrectFunctionInput;
+import com.example.multimathsolver.domain.Repository;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -14,12 +15,12 @@ import java.util.Collection;
 
 @RunWith(Parameterized.class)
 public class DeadLockedDNFTest {
-
+    Repository repository = new RepositoryImpl();
     BooleanFunction function;
     String expectedDeadLockedDNF;
 
     public DeadLockedDNFTest(String expression, String expectedDeadLockedDNF) throws IncorrectFunctionInput {
-        this.function = new BooleanFunction(new StringExpressionHandler(expression));
+        this.function = repository.getBooleanFunction(expression);
         this.expectedDeadLockedDNF = expectedDeadLockedDNF;
     }
     @Parameterized.Parameters
@@ -47,6 +48,6 @@ public class DeadLockedDNFTest {
 
     @Test
     public void testDeadLockedDNF() {
-        Assert.assertEquals(expectedDeadLockedDNF, function.getDeadLockedDNF());
+        Assert.assertEquals(expectedDeadLockedDNF, repository.getDeadLockedDNF(function));
     }
 }
