+++ app/src/test/java/com/example/multimathsolver/booleanalgebra/PerfectDNFTest.java
@@ -1,8 +1,9 @@
 package com.example.multimathsolver.booleanalgebra;
 
-import com.example.multimathsolver.data.booleanalgebra.BooleanFunction;
-import com.example.multimathsolver.data.booleanalgebra.IncorrectFunctionInput;
-import com.example.multimathsolver.data.booleanalgebra.StringExpressionHandler;
+import com.example.multimathsolver.data.RepositoryImpl;
+import com.example.multimathsolver.domain.BooleanFunction;
+import com.example.multimathsolver.domain.IncorrectFunctionInput;
+import com.example.multimathsolver.domain.Repository;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -14,12 +15,12 @@ import java.util.Collection;
 
 @RunWith(Parameterized.class)
 public class PerfectDNFTest {
-
+    Repository repository = new RepositoryImpl();
     BooleanFunction function;
     String expectedPerfectDNF;
 
     public PerfectDNFTest(String expression, String expectedPerfectDNF) throws IncorrectFunctionInput {
-        this.function = new BooleanFunction(new StringExpressionHandler(expression));
+        this.function = repository.getBooleanFunction(expression);
         this.expectedPerfectDNF = expectedPerfectDNF;
     }
     @Parameterized.Parameters
@@ -34,6 +35,6 @@ public class PerfectDNFTest {
 
     @Test
     public void testPerfectDNF() {
-        Assert.assertEquals(expectedPerfectDNF, function.getPerfectDNF());
+        Assert.assertEquals(expectedPerfectDNF, repository.getPerfectDNF(function));
     }
 }
