+++ app/src/test/java/com/example/multimathsolver/booleanalgebra/TableTest.java
@@ -1,8 +1,9 @@
 package com.example.multimathsolver.booleanalgebra;
 
-import com.example.multimathsolver.data.booleanalgebra.BooleanFunction;
-import com.example.multimathsolver.data.booleanalgebra.IncorrectFunctionInput;
-import com.example.multimathsolver.data.booleanalgebra.StringExpressionHandler;
+import com.example.multimathsolver.data.RepositoryImpl;
+import com.example.multimathsolver.domain.BooleanFunction;
+import com.example.multimathsolver.domain.IncorrectFunctionInput;
+import com.example.multimathsolver.domain.Repository;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -14,11 +15,12 @@ import java.util.Collection;
 
 @RunWith(Parameterized.class)
 public class TableTest {
+    Repository repository = new RepositoryImpl();
     private final BooleanFunction function;
     private final int[] expectedTableResult;
 
     public TableTest(String expression, int[] list) throws IncorrectFunctionInput {
-        this.function = new BooleanFunction(new StringExpressionHandler(expression));
+        this.function = repository.getBooleanFunction(expression);
         this.expectedTableResult = list;
     }
